% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig1a_functions.R
\name{subset_data}
\alias{subset_data}
\title{Subset data}
\usage{
subset_data(df, genera)
}
\arguments{
\item{df}{A data frame, data frame extension (e.g. a tibble), or a lazy
data frame (e.g. from dbplyr or dtplyr). Requires column \code{Genus}.}

\item{genera}{A character vector that corresponds to the genera to be
dropped.}
}
\value{
An object of the same type as \code{df}. Properties of the output:
\itemize{
\item Retained rows appear in the same order as in the input
\item Columns and data frame attributes are preserved
}
}
\description{
\code{subset_data} drops observations according to their genera.
}
\examples{
subset_data(data, c("Elephantulus", "Orcinus"))

#' \dontrun{
subset_data("a")
}
}
